refactor(cli): split auth command into submodules [OPE-265]
